# docs/src/openapi.yaml
...
openapi: 3.1.0
info:
  title: CSM HTTP API
  version: "1.0"
  description: >-
    Continuous Security Monitor daemon API. API clients authenticate with a
    bearer token from webui.tokens. Read-scope tokens authorize the GET
    endpoints marked x-csm-scope=read; management routes require an
    admin-scope token. Every failure is a non-2xx status with an Error body.
    A state change answers an ActionResult, 202 for work that continues after
    the answer. A list answers an ItemList. Times are RFC 3339 instants in
    UTC, durations are seconds in keys ending in _seconds, severities are
    CRITICAL, HIGH or WARNING, and an empty list is [] rather than null.
servers:
  - url: https://localhost:9443
    description: CSM daemon HTTPS listener (operator-configured host/port)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        webui.tokens entry. scope=read authorizes read endpoints; scope=admin
        authorizes all endpoints.
  responses:
    Unauthorized:
      description: Missing, invalid, or insufficient-scope token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: CSRF validation failed or cross-origin API request blocked.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    JSONObject:
      type: object
      description: Endpoint-specific JSON. Per-field schema is a documented follow-up.
      additionalProperties: true
    Error:
      type: object
      description: Every failure. Some routes add detail next to error.
      required: [error]
      properties:
        error: { type: string }
      additionalProperties: true
...
security:
  - bearerAuth: []
paths:
  "/api/v1/account":
    get:
      operationId: get_account
      summary: "get /api/v1/account (scope: admin)"
      x-csm-scope: admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JSONObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  "/api/v1/accounts":
    get:
      operationId: get_accounts
      summary: "get /api/v1/accounts (scope: admin)"
      x-csm-scope: admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JSONObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  "/api/v1/email/held":
    get:
      operationId: get_email_held
      summary: "get /api/v1/email/held (scope: admin)"
      x-csm-scope: admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JSONObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  "/api/v1/email/held/":
    post:
      operationId: post_email_held_
      summary: "post /api/v1/email/held/ (scope: admin)"
      x-csm-scope: admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JSONObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  "/api/v1/email/quarantine":
    get:
      operationId: get_email_quarantine
      summary: "get /api/v1/email/quarantine (scope: admin)"
      x-csm-scope: admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JSONObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  "/api/v1/email/quarantine/":
    post:
      operationId: post_email_quarantine_
      summary: "post /api/v1/email/quarantine/ (scope: admin)"
      x-csm-scope: admin
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JSONObject" }
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JSONObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  "/api/v1/findings":
    get:
      operationId: get_findings
      summary: "get /api/v1/findings (scope: read)"
      x-csm-scope: read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JSONObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...